+++ kubernetes-tests/src/test/java/io/fabric8/kubernetes/client/mock/CSIStorageCapacityTest.java
@@ -38,7 +38,7 @@ class CSIStorageCapacityTest {
 
   @Test
   void load() {
-    List<HasMetadata> items = client.load(getClass().getResourceAsStream("/test-csistoragecapacity.yml")).get();
+    List<HasMetadata> items = client.load(getClass().getResourceAsStream("/test-csistoragecapacity.yml")).items();
     assertThat(items).isNotNull();
     AssertionsForClassTypes.assertThat(items.get(0))
         .isInstanceOf(CSIStorageCapacity.class)