+++ kubernetes-tests/src/test/java/io/fabric8/kubernetes/client/mock/V1PriorityClassTest.java
@@ -168,7 +168,8 @@ class V1PriorityClassTest {
 
   @Test
   void testLoadFromFile() {
-    assertNotNull(client.scheduling().priorityClass().load(getClass().getResourceAsStream("/test-v1-priorityclass.yml")).get());
+    assertNotNull(
+        client.scheduling().priorityClass().load(getClass().getResourceAsStream("/test-v1-priorityclass.yml")).item());
   }
 
 }