+++ kubernetes-tests/src/test/java/io/fabric8/kubernetes/client/mock/V1beta2PriorityLevelConfigurationTest.java
@@ -38,7 +38,7 @@ class V1beta2PriorityLevelConfigurationTest {
 
   @Test
   void load() {
-    List<HasMetadata> items = client.load(getClass().getResourceAsStream("/v1beta2-prioritylevelconfiguration.yml")).get();
+    List<HasMetadata> items = client.load(getClass().getResourceAsStream("/v1beta2-prioritylevelconfiguration.yml")).items();
     assertThat(items).isNotNull().hasSize(1);
     AssertionsForClassTypes.assertThat(items.get(0))
         .isInstanceOf(PriorityLevelConfiguration.class)